Add About page, placeholder logo, general styling
A lot of stuff here is mind-dump filler text, as well as a hastily found
logo that is free(?), TODO credit is in NOTES.md file. General style is
good, still worried about the footer, which can hike if the content
doesn't fill available viewport height. CSS for that is fiddly, it's in
the backlog (I think). Dead routes are on next sprint